collect content parallel
